Added tests

+++ config/initializers/yandex_newsletter.rb
@@ -1,4 +1,4 @@
-unless Rails.env.test?
+if !Rails.env.test? && Que.execute("SELECT * FROM information_schema.tables WHERE table_name = 'que_jobs'")
   Que.execute 'DELETE FROM que_jobs'
   YandexNewsPuller.enqueue
 end
